backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy==2.0.36

# Validation (included with FastAPI, but pinned for clarity)
pydantic==2.10.3

# Testing
pytest==8.3.4
httpx==0.28.1

# Authentication
python-jose[cryptography]==3.5.0
bcrypt==5.0.0

# Utilities
python-dotenv==1.0.1
